rename cipher->ciphers
